feat(validation): add validation and enable add-ons
auto-validation and error message can be set through the properties:
`isAutoValidating` and `errorMessage`.

Auto-validation is triggered on blur or on input. Validation criteria
can be set with `isRequired` or `pattern`. `pattern` takes a regex
as string, e.g., `[a-zA-Z]*`.

`errorMessage` is only available with `molecule-input`.

It is possible to inject add-ons in `molecule-input-container` through
`addOns` property. Add-ons are function references in an array, e.g.:

addOns: [moleculeInputError]

Add-ons are instantiated automatically with DOM and current state as properties.
If you need to set special values for an add-on, set them with `bind()`.

Add-ons are rendered to the view as the last section.

BREAKING CHANGE:

`shared/renderLabel` has a new signature.

Before: `renderLabel(dialogueName, label)`
After: `renderLabel({dialogueName, label})`

Closes #10